+++ src/schema/artist.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Artist {
-    pub id: i64,
+    pub id: u32,
     pub name: String,
 }
